requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
drf-extra-fields==3.7.0
filetype==1.2.0
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
PyPDF2==3.0.1
pytest==8.3.4
sqlparse==0.5.3
